HW1-naver_FAQ_chatbot/Minhye/chatbot/pyproject.toml
[tool.poetry]
name = "chatbot" 
version = "0.0.1"
description = ""
authors = ["Minhye"]
packages = [{include = "chatbot"}]

[tool.poetry.dependencies]
python = "^3.9" 
openai = "^1.8.0"
pandas = "^2.2.0"
scikit-learn = "^1.4.2"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
black = {version = "^23.1.0", allow-prereleases = true}
ipykernel = "^6.21.1"
pytest = "7.2.0"
ruff = "^0.0.257"
pytest-asyncio = "^0.21.0"

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
